cAER: DAVIS: abstract frame initialization, and IMU gyro/accel scale calculation. Use them where needed. Recalculate IMU scales only on special event that specifies them. Use calculated values when putting the IMU values from device into packet.
git-svn-id: svn://svn.code.sf.net/p/jaer/code@5884 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ cAER/trunk/modules/ini/davis_common.h
@@ -81,8 +81,8 @@ struct davisCommon_state {
 	// IMU specific fields
 	uint8_t imuCount;
 	uint8_t imuTmpData;
-	uint8_t imuAccelScale;
-	uint8_t imuGyroScale;
+	float imuAccelScale;
+	float imuGyroScale;
 	// Polarity Packet State
 	caerPolarityEventPacket currentPolarityPacket;
 	uint32_t currentPolarityPacketPosition;
